services/pdf-parser/pyproject.toml
[project]
name = "moneypulse-pdf-parser"
version = "1.0.0"
description = "MoneyPulse PDF Parser Microservice"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pdfplumber>=0.11.0",
    "tabula-py>=2.9.0",
    "httpx>=0.28.0",
    "pydantic>=2.10.0",
    "python-multipart>=0.0.18",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
    "fpdf2>=2.8.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
